LayoutTests/svg/animations/svglengthlist-animation-2-expected.txt

SVG 1.1 dynamic animation tests

ABCD
Test 'to' animation of SVGLengthList with different LengthTypes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS text.x.animVal.numberOfItems is 4
PASS text.x.animVal.getItem(0).value is 50
PASS text.x.animVal.getItem(1).value is 70
PASS text.x.animVal.getItem(2).value is 90
PASS text.x.animVal.getItem(3).value is 110
PASS text.x.baseVal.numberOfItems is 4
PASS text.x.baseVal.getItem(0).value is 50
PASS text.x.baseVal.getItem(1).value is 70
PASS text.x.baseVal.getItem(2).value is 90
PASS text.x.baseVal.getItem(3).value is 110
PASS text.x.animVal.numberOfItems is 4
PASS text.x.animVal.getItem(0).value is 81.7
PASS text.x.animVal.getItem(1).value is 75
PASS text.x.animVal.getItem(2).value is 144.5
PASS text.x.animVal.getItem(3).value is 197.3
PASS text.x.baseVal.numberOfItems is 4
PASS text.x.baseVal.getItem(0).value is 50
PASS text.x.baseVal.getItem(1).value is 70
PASS text.x.baseVal.getItem(2).value is 90
PASS text.x.baseVal.getItem(3).value is 110
PASS text.x.animVal.numberOfItems is 4
PASS text.x.animVal.getItem(0).value is 113.4
PASS text.x.animVal.getItem(1).value is 80
PASS text.x.animVal.getItem(2).value is 100
PASS text.x.animVal.getItem(3).value is 384
PASS text.x.baseVal.numberOfItems is 4
PASS text.x.baseVal.getItem(0).value is 50
PASS text.x.baseVal.getItem(1).value is 70
PASS text.x.baseVal.getItem(2).value is 90
PASS text.x.baseVal.getItem(3).value is 110
PASS text.x.animVal.numberOfItems is 4
PASS text.x.animVal.getItem(0).value is 50
PASS text.x.animVal.getItem(1).value is 70
PASS text.x.animVal.getItem(2).value is 90
PASS text.x.animVal.getItem(3).value is 110
PASS text.x.baseVal.numberOfItems is 4
PASS text.x.baseVal.getItem(0).value is 50
PASS text.x.baseVal.getItem(1).value is 70
PASS text.x.baseVal.getItem(2).value is 90
PASS text.x.baseVal.getItem(3).value is 110
PASS successfullyParsed is true

TEST COMPLETE
